Major changes to ax_lib_samtools.m4.
It seems that the installation paths for the samtools headers and libbam.a
haven't been standardised, so each distribution may put those files in
different locations.

Moreover, the check for libbam.a needs also -lz and -lpthread to compile.

This change allows to find the path when using Ubuntu, where compilation is
successful if configure is called with option "--with-samtools=/usr".

Since mendel-gpu cannon build without samtools, now the script will stop
with an error if the user tries to disable samtools. Similarly, warnings
are raised if --with-samtools is specified but a valid path is not given.